Fixed compilation and modernized CMake
compilation was failing due to the header file `backward.hpp` not being
found in AAClient compile, this is fixed by using
`add_backward(AAClient)` just like in AAServer.

Instead of using a Git Submodule, we can use a modernized CMake
equivalent to integrate this further into CMake, allowing for a version
of Backward installed to the system to be used if it is found, and if
not, clone it and compile it from source (like the gitsubmodule would do).
This allows seamless transitions between system installed software and
not.

There was also a minor issue with the include for std::set not being
included, and that was fixed by adding `#include <set>` in AAServer
InputChannelHandler.h