Attempted to fix runtime errors for `read()` error EOVERFLOW
The first `read()` in the `handleSwitchToAccessoryMode()` function
in AAServer/src/ModeSwitcher.cpp fails at some point with The
`errno` being 75, or EOVERFLOW. This error is described as
"Value too large to be stored in data type" (https://man7.org/linux/man-pages/man3/errno.3.html)
This system error message is reported when a file operation is performed
on a file using a non-large file data type. (Anything less than 2Gb on a
32bit system)

To fix this issue, we can define the `_FILE_OFFSET_BITS` to the value 64
to change the file offset type `off_t` from a 32 bit type to a 64 bit
type. Changing the build system to effectively do this does change that
data type (verified in GDB on the target), yet, this does not fix the
issue. Some articles mention that the low leverl `read` family functions
use `size_t` as their file offsets. This type however, is limited to a
32 bit wide value, and can be even narrower. It might be possible to try
to use the C standard (or even C++ standard) library file operation
functions to see if that helps.

Along with those issues, the `checkError` function used the
`std::to_string` function to return the errno, when it could use
`std::strerror` to convert the errno to a human readable format.